recollapse all of it
