+++ pkg/controller/issueapi/controller.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package issueapi implements the API handler for taking a code request, assigning
+// package issueapi implements the API handler for taking a code request, assigning
 // an OTP, saving it to the database and returning the result.
 // This is invoked over AJAX from the Web frontend.
 package issueapi
