+++ pkg/controller/issueapi/metrics.go
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package issuemetric implements the metrics for code issuance
-package issuemetric
+package issueapi
 
 import (
 	enobservability "github.com/google/exposure-notifications-server/pkg/observability"
@@ -27,37 +26,37 @@ import (
 const metricPrefix = observability.MetricRoot + "/api/issue"
 
 var (
-	RealmTokenUsed = stats.Int64(metricPrefix+"/realm_token_used", "# of realm token used.", stats.UnitDimensionless)
-	SMSLatencyMs   = stats.Float64(metricPrefix+"/sms_request", "# of sms requests", stats.UnitMilliseconds)
-	LatencyMs      = stats.Float64(metricPrefix+"/request", "# of code issue requests", stats.UnitMilliseconds)
+	mRealmTokenUsed = stats.Int64(metricPrefix+"/realm_token_used", "# of realm token used.", stats.UnitDimensionless)
+	mSMSLatencyMs   = stats.Float64(metricPrefix+"/sms_request", "# of sms requests", stats.UnitMilliseconds)
+	mLatencyMs      = stats.Float64(metricPrefix+"/request", "# of code issue requests", stats.UnitMilliseconds)
 )
 
 func init() {
 	enobservability.CollectViews([]*view.View{
 		{
 			Name:        metricPrefix + "/request_count",
-			Measure:     LatencyMs,
+			Measure:     mLatencyMs,
 			Description: "Count of code issue requests",
 			TagKeys:     observability.APITagKeys(),
 			Aggregation: view.Count(),
 		},
 		{
 			Name:        metricPrefix + "/request_latency",
-			Measure:     LatencyMs,
+			Measure:     mLatencyMs,
 			Description: "The latency distribution of code issue requests",
 			TagKeys:     observability.APITagKeys(),
 			Aggregation: ochttp.DefaultLatencyDistribution,
 		},
 		{
 			Name:        metricPrefix + "/sms_request_count",
-			Measure:     SMSLatencyMs,
+			Measure:     mSMSLatencyMs,
 			Description: "The # of SMS requests",
 			TagKeys:     append(observability.CommonTagKeys(), observability.ResultTagKey),
 			Aggregation: view.Count(),
 		},
 		{
 			Name:        metricPrefix + "/sms_request_latency",
-			Measure:     SMSLatencyMs,
+			Measure:     mSMSLatencyMs,
 			Description: "The # of SMS requests",
 			TagKeys:     append(observability.CommonTagKeys(), observability.ResultTagKey),
 			Aggregation: ochttp.DefaultLatencyDistribution,
@@ -66,7 +65,7 @@ func init() {
 			Name:        metricPrefix + "/realm_token_used_count",
 			Description: "The count of # of realm token used.",
 			TagKeys:     observability.CommonTagKeys(),
-			Measure:     RealmTokenUsed,
+			Measure:     mRealmTokenUsed,
 			Aggregation: view.Count(),
 		},
 	}...)
